refactor: 统一错误处理并拆分 command/preview 模块
- 新增 error.rs，引入 thiserror 定义统一错误类型 QuickLookError
- 拆分 command.rs → commands/{archive,document,image,audio,video,system}.rs
- 拆分 preview.rs → preview/{hook,route,window}.rs
- 所有命令函数和 helper 函数改用 QuickLookError 替代 String 错误
- 为 QuickLookError 实现 Serialize 以兼容 Tauri command 序列化
