+++ src-tauri/Cargo.toml
@@ -22,6 +22,7 @@ tauri-build = { version = "2.6.1", features = [] }
 serde_json = "1.0"
 serde = { version = "1.0", features = ["derive"] }
 log = "0.4"
+thiserror = "2"
 tauri = { version = "2.11.1", features = [
     "protocol-asset",
     "unstable",
