--- src-tauri/src/command.rs
@@ -1,194 +1,0 @@
-use log::{set_max_level, LevelFilter};
-use quicklook_archive::{extractors, Extract};
-use quicklook_docs as docs;
-use std::path::PathBuf;
-use tauri::{command, AppHandle, Manager};
-use windows::Win32::Foundation::HWND;
-
-use crate::helper::{audio, ffmp, image as image_helper, monitor, win};
-
-#[command]
-pub fn check_ffmpeg() -> bool {
-    ffmp::check_ffmpeg()
-}
-
-#[command]
-pub async fn convert_video_to_hls(path: String) -> Result<String, String> {
-    tauri::async_runtime::spawn_blocking(move || ffmp::convert_video_to_hls(&path))
-        .await
-        .map_err(|e| format!("转码任务执行失败: {}", e))?
-}
-
-#[command]
-pub fn cancel_video_conversion() {
-    ffmp::cancel_video_conversion()
-}
-
-#[command]
-pub fn show_open_with_dialog(app: AppHandle, path: &str) {
-    if let Some(preview_window) = app.get_webview_window("preview") {
-        let hwnd = preview_window.hwnd().map_or(HWND::default(), |hwnd| hwnd);
-        let _ = win::show_open_with_dialog(path, hwnd);
-    }
-}
-
-#[command]
-pub fn archive(path: &str, mode: &str) -> Result<Vec<Extract>, String> {
-    log::info!("开始处理压缩文件: {}, 扩展名: {}", path, mode);
-    let result = match mode {
-        "zip" => extractors::zip::zip_extract(path).map_err(|e| e.to_string()),
-        "tar" => extractors::tar::list_tar_entries(path).map_err(|e| e.to_string()),
-        "gz" | "tgz" => extractors::tar::list_tar_gz_entries(path).map_err(|e| e.to_string()),
-        "bz2" | "tbz2" => extractors::tar::list_tar_bz2_entries(path).map_err(|e| e.to_string()),
-        "xz" | "txz" => extractors::tar::list_tar_xz_entries(path).map_err(|e| e.to_string()),
-        "zst" | "tzst" => extractors::zst::list_tar_zst_entries(path).map_err(|e| e.to_string()),
-        "7z" => extractors::sevenz::list_7z_entries(path).map_err(|e| e.to_string()),
-        "cpio" => extractors::cpio::list_cpio_entries(path).map_err(|e| e.to_string()),
-        "ar" | "deb" | "a" => extractors::ar::list_ar_entries(path).map_err(|e| e.to_string()),
-        "jar" | "war" | "ear" | "apk" | "aar" | "whl" | "vsix" | "nupkg"
-        | "crx" | "xpi" | "egg" | "kra" | "xps" | "oxps" => {
-            extractors::zip::zip_extract(path).map_err(|e| e.to_string())
-        },
-        _ => Err("不支持的压缩格式".to_string()),
-    };
-
-    match &result {
-        Ok(entries) => {
-            log::info!("成功处理压缩文件，共{}个条目", entries.len());
-        },
-        Err(e) => {
-            log::error!("压缩文件处理失败: {}", e);
-        },
-    }
-
-    result
-}
-
-#[command]
-pub fn document(path: &str, mode: &str) -> Result<docs::Docs, String> {
-    match mode {
-        "csv" => docs::Docs::csv(path).map_err(|e| e.to_string()),
-        "xlsx" | "xls" | "xlsm" | "xlsb" | "xla" | "xlam" | "ods" => {
-            docs::Docs::excel(path).map_err(|e| e.to_string())
-        },
-        "docx" => docs::Docs::docx(path).map_err(|e| e.to_string()),
-        _ => Err("Not Support".to_string()),
-    }
-}
-
-#[command]
-pub fn get_monitor_info() -> monitor::MonitorInfo {
-    monitor::get_monitor_info()
-}
-
-#[command]
-pub fn get_default_program_name(path: &str) -> Result<String, String> {
-    win::get_default_program_name(path)
-}
-
-#[command]
-pub fn set_log_level(level: usize) -> Result<(), String> {
-    let level_filter = match level {
-        1 => LevelFilter::Error,
-        2 => LevelFilter::Warn,
-        3 => LevelFilter::Info,
-        4 => LevelFilter::Debug,
-        5 => LevelFilter::Trace,
-        _ => LevelFilter::Off,
-    };
-    set_max_level(level_filter);
-    Ok(())
-}
-
-/// 将 PSD、HEIC/HEIF 等图片格式转换为 PNG 并缓存。
-/// 返回转换后 PNG 文件的路径。
-#[command]
-pub fn convert_to_png(path: &str) -> Result<String, String> {
-    use std::collections::hash_map::DefaultHasher;
-    use std::hash::{Hash, Hasher};
-
-    let mut hasher = DefaultHasher::new();
-    path.hash(&mut hasher);
-    if let Ok(meta) = std::fs::metadata(path) {
-        if let Ok(modified) = meta.modified() {
-            modified.hash(&mut hasher);
-        }
-    }
-    let hash = hasher.finish();
-
-    let mut images_dir: PathBuf = std::env::temp_dir();
-    images_dir.push("quicklook_images");
-    std::fs::create_dir_all(&images_dir).map_err(|e| e.to_string())?;
-
-    let mut temp_path = images_dir;
-    temp_path.push(format!("quicklook_{:x}.png", hash));
-
-    if temp_path.exists() {
-        log::info!("命中图片缓存: {:?}", temp_path);
-        return Ok(temp_path.to_string_lossy().to_string());
-    }
-
-    let ext = std::path::Path::new(path)
-        .extension()
-        .and_then(|e| e.to_str())
-        .unwrap_or("")
-        .to_lowercase();
-
-    match ext.as_str() {
-        "psd" => image_helper::psd_to_png(path, &temp_path)?,
-        "heic" | "heif" => image_helper::heic_to_png(path, &temp_path)?,
-        _ => image_helper::image_to_png(path, &temp_path)?,
-    }
-
-    Ok(temp_path.to_string_lossy().to_string())
-}
-
-#[command]
-pub fn read_audio_info(path: &str) -> Option<audio::MusicInfo> {
-    audio::read_music_info(path)
-}
-
-#[command]
-pub fn parse_lrc(path: &str) -> Result<audio::Lrc, String> {
-    audio::parse_lrc(path)
-}
-
-/// 清理图片转码缓存目录。
-/// 返回被删除的文件数量。
-#[command]
-pub fn clear_image_cache() -> Result<u32, String> {
-    let images_dir = std::env::temp_dir().join("quicklook_images");
-    if !images_dir.exists() {
-        log::info!("quicklook_images 目录不存在，无需清理");
-        return Ok(0);
-    }
-    let entries = std::fs::read_dir(&images_dir).map_err(|e| e.to_string())?;
-
-    let mut removed = 0u32;
-    for entry in entries.flatten() {
-        if entry.path().is_file() {
-            match std::fs::remove_file(entry.path()) {
-                Ok(_) => {
-                    removed += 1;
-                    log::info!("已清理图片缓存: {}", entry.path().display());
-                },
-                Err(e) => {
-                    log::warn!("清理图片缓存失败: {}, 错误: {}", entry.path().display(), e);
-                },
-            }
-        }
-    }
-    log::info!("共清理 {} 个图片缓存文件", removed);
-    Ok(removed)
-}
-
-/// 汇总清理所有 quicklook 产生的缓存，包含 ffmpeg HLS 转码缓存和图片转码缓存。
-/// 返回被删除的目录/文件总数量。
-#[command]
-pub fn clear_cache() -> Result<u32, String> {
-    let mut total = 0u32;
-    total += ffmp::clear_ffmpeg_cache()?;
-    total += clear_image_cache()?;
-    log::info!("缓存清理完成，共删除 {} 个目录/文件", total);
-    Ok(total)
-}
